rmtree cache in extractor

+++ lib/py/etl/paradicms_etl/extractors/airtable_extractor.py
@@ -52,8 +52,9 @@ class AirtableExtractor:
         base = self.__extract_base_metadata(base_id=self.__base_id, force=force)
 
         records_by_table = {}
-        rmtree(self.__cache_dir_path, ignore_errors=True)
-        self.__cache_dir_path.mkdir(parents=True)
+        if force:
+            rmtree(self.__cache_dir_path, ignore_errors=True)
+        self.__cache_dir_path.mkdir(parents=True, exist_ok=True)
         for table, query_parameters in self.__tables.items():
             records_by_table[table] = self.__extract_table_records(
                 force=force, table=table, query_parameters=query_parameters
